+++ LigidPainter/Main/src/Core/UiActions.cpp
@@ -205,7 +205,7 @@ UiActionsData UiActions::uiActions(GLFWwindow* window ,CallbckData callbackData,
     uiData.textureDemonstratorButtonPressed = textureDemonstratorButtonPressed;
     return uiData;
 }
-bool UiActions::updateRangeValues(GLFWwindow* window, float xOffset,float yOffset, int screenWidth, int screenHeight){
+bool UiActions::updateRangeValues(GLFWwindow* window, double xOffset,double yOffset, int screenWidth, int screenHeight){
     bool hideCursor;
     LigidPainter ligid;
