+++ LigidPainter/Main/src/Core/UserInterface.h
@@ -3,7 +3,7 @@
 
 struct ColorData //LigidPainter color palette
 {
-	glm::vec3 LigidPainterThemeColor = glm::vec3(0.043f,0.635f,0.823);
+	glm::vec3 LigidPainterThemeColor = glm::vec3(0.043f,0.635f,0.823f);
 
 	glm::vec3 viewportBackColor = glm::vec3(0.23f, 0.23f, 0.23f);
 	glm::vec3 panelColor = glm::vec3(0.15f, 0.15f, 0.15f);
@@ -45,15 +45,15 @@ public:
 
 	void box(float width, float height, float position_x, float position_y, std::string text, glm::vec3 color, float textRatio, bool isTextBox, bool isMaskImageBox, float buttonCurveReduce,float z,glm::vec3 colorTransitionColor , float mixVal);
 	void iconBox(float width, float height, float position_x, float position_y,float z, unsigned int icon,float mixVal,glm::vec3 color,glm::vec3 colorHover);
-	bool isMouseOnButton(GLFWwindow* window, float width, float height, float position_x, float position_y, int mouseXpos, int mouseYpos, bool isPanelMoving);
-	bool isMouseOnCoords(GLFWwindow*window,int mouseXpos, int mouseYpos,std::vector<float> buttonCoor,bool isPanelMoving);
+	bool isMouseOnButton(GLFWwindow* window, float width, float height, float position_x, float position_y, double mouseXpos, double mouseYpos, bool isPanelMoving);
+	bool isMouseOnCoords(GLFWwindow*window,double mouseXpos, double mouseYpos,std::vector<float> buttonCoor,bool isPanelMoving);
 	void setViewportBgColor();
 	void renderText(unsigned int program, std::string text, float x, float y, float scale);
 	void uploadChars();
 	void renderMenubar(GLFWwindow* window);
 	void checkBox(float position_x, float position_y, std::string text, glm::vec3 color, bool mouseHover, bool checked);
 	void panelChangeButton(float position_x, float position_y);
-	bool isMouseOnPanelChangeButton(GLFWwindow* window, float position_x, float position_y, int mouseXpos, int mouseYpos);
+	bool isMouseOnPanelChangeButton(GLFWwindow* window, float position_x, float position_y, double mouseXpos, double mouseYpos);
 	void rangeBar(float position_x, float position_y, float value);
 	void colorBox(float position_x, float position_y, float valueX, float valueY);
 	glm::vec3 colorRect(float position_x, float position_y,float value, unsigned int FBO,GLFWwindow* window,glm::mat4 projection,bool updateHueVal);
